feat: add collect ticket by external id
